Add logic for processing Tasks

+++ src/main/kotlin/com/io/telegram/TelegramResult.kt
@@ -2,5 +2,5 @@ package com.io.telegram
 
 data class TelegramResult(
     val behaviour: TelegramBehaviour,
-    val doFinish: suspend (messageIds: Pair<Int, String>) -> Unit
+    val doFinish: suspend (messageIds: Pair<Long, String>) -> Unit
 )
